[skipruntime] Replace `playwright` with `mocha` for tests.
The only reason we were using Playwright was so that we could run
tests within browsers. While this made sense for skdb, it is not
expected to be common for the skipruntime server to run in a browser.

Moreover, it turns out the tests were not actually being executed in
the browsers themselves, but in the nodejs environments spawning the
browsers, meaning we were getting exactly zero benefit from using
playwright.

This commit changes the test framework from Playwright to mochajs +
earl, which simplifies and speeds things up. Moreover, this obviates
the need for a separate `package.json` for tests.